requirements.txt
opencv-python==4.10.0.84
Pillow==10.0.1
